# wpf-rest-client-aspnetcore/RestWebApiServer/RestWebApiServer/Controllers/BooksController.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc; // [Route], [ApiController], [HttpGet]
using Microsoft.EntityFrameworkCore;
using RestWebApiServer.Data;
using RestWebApiServer.Models;
using Swashbuckle.AspNetCore.Annotations; // 注意! これだけでは有効にならない.
                                          // See Startup#ConfigureServices()

namespace RestWebApiServer.Controllers
{

// [Route("api/[controller]")]  これだと api/Books と大文字になる
[Route("api/books")]
[ApiController]
public class BooksController : ControllerBase
{
    readonly RestWebApiServerContext _context;

    public BooksController(RestWebApiServerContext context)
    {
        _context = context;
    }

    // GET: api/Books
    // Client クラスに押し込められるので, FindAllだけでは分からない。
    [HttpGet]
    [SwaggerOperation(OperationId = "FindAllBooks")] 
    public async Task<ActionResult<IEnumerable<Book>>> GetBook()
    {
        return await _context.Book.ToListAsync();
    }

        // GET: api/Books/5
    [HttpGet("{id}")]
    [SwaggerOperation(OperationId = "FindBook")]
    public async Task<ActionResult<Book>> GetBook(int id)
    {
        var book = await _context.Book.FindAsync(id);

            if (book == null)
            {
                return NotFound();
            }

            return book;
        }

        // PUT: api/Books/5
        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
    [HttpPut("{id}")]
    [SwaggerOperation(OperationId = "UpdateBook")]
    public async Task<IActionResult> PutBook(int id, [FromBody] Book book)
    {
            if (id != book.Id)
            {
                return BadRequest();
            }

            _context.Entry(book).State = EntityState.Modified;

            try
            {
                await _context.SaveChangesAsync();
            }
            catch (DbUpdateConcurrencyException)
            {
                if (!BookExists(id))
                {
                    return NotFound();
                }
                else
                {
                    throw;
                }
            }

            return NoContent();
        }

        // POST: api/Books
        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
    [HttpPost]
    [SwaggerOperation(OperationId = "CreateBook")]
    public async Task<ActionResult<Book>> PostBook([FromBody] Book book)
    {
            _context.Book.Add(book);
            await _context.SaveChangesAsync();

            return CreatedAtAction("GetBook", new { id = book.Id }, book);
        }

        // DELETE: api/Books/5
    [HttpDelete("{id}")]
    [SwaggerOperation(OperationId = "DestroyBook")]
    public async Task<IActionResult> DeleteBook(int id)
    {
            var book = await _context.Book.FindAsync(id);
            if (book == null)
            {
                return NotFound();
            }

            _context.Book.Remove(book);
            await _context.SaveChangesAsync();

            return NoContent();
        }

        private bool BookExists(int id)
        {
            return _context.Book.Any(e => e.Id == id);
        }
    }
}