# wpf-rest-client-aspnetcore/WPF Client Web API-renew/WPF Client Web API/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RestWebApiServer",
    "version": "v1"
  },
  "paths": {
    "/api/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "operationId": "FindAllBooks",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "operationId": "CreateBook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{id}": {
      "get": {
        "tags": [
          "Books"
        ],
        "operationId": "FindBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Books"
        ],
        "operationId": "UpdateBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Books"
        ],
        "operationId": "DestroyBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/creators": {
      "get": {
        "tags": [
          "Creators"
        ],
        "operationId": "FindAllCreators",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Creator"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Creator"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Creator"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Creators"
        ],
        "operationId": "CreateCreator",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Creator"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Creator"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Creator"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Creator"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Creator"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Creator"
                }
              }
            }
          }
        }
      }
    },
    "/api/creators/{id}": {
      "get": {
        "tags": [
          "Creators"
        ],
        "operationId": "FindCreator",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Creator"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Creator"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Creator"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Creators"
        ],
        "operationId": "UpdateCreator",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Creator"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Creator"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Creator"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Creators"
        ],
        "operationId": "DestroyCreator",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthorBook": {
        "required": [
          "bookId",
          "creatorId",
          "sort",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "creatorId": {
            "type": "integer",
            "format": "int32"
          },
          "bookId": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/Type"
          },
          "sort": {
            "type": "integer",
            "format": "int32"
          },
          "creator": {
            "$ref": "#/components/schemas/Creator"
          },
          "book": {
            "$ref": "#/components/schemas/Book"
          }
        },
        "additionalProperties": false
      },
      "Book": {
        "required": [
          "createdAt",
          "description",
          "lockVersion",
          "title",
          "updatedAt",
          "year"
        ],
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lockVersion": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "title": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "authors_books": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorBook"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Creator": {
        "required": [
          "createdAt",
          "lockVersion",
          "name",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lockVersion": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "name": {
            "type": "string"
          },
          "authors_books": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorBook"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Type": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "WeatherForecast": {
        "required": [
          "summary"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}